+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+ ### Added
+ - Add support for ordering material requirements
+
 ## [0.2.1] - 2021-10-18
 
 ### Fixed